[common/metatypes] refactor: move datatypes used by meta into common/metatypes, so that other crates are free to use them
