+++ common/metatypes/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2021"
 
 [dependencies]
 serde = { version = "1.0", features = ["derive"] }
+async-raft = { git = "https://github.com/datafuse-extras/async-raft", tag = "v0.6.2-alpha.14" }
 
 [dev-dependencies]
 pretty_assertions = "0.7"
