+++ common/metatypes/src/lib.rs
@@ -21,14 +21,29 @@ use std::fmt;
 use std::fmt::Debug;
 use std::fmt::Formatter;
 
+pub use cluster::Node;
+pub use cluster::Slot;
+pub use cmd::Cmd;
 pub use errors::ConflictSeq;
+pub use log_entry::LogEntry;
 pub use match_seq::MatchSeq;
 pub use match_seq::MatchSeqExt;
+pub use raft_txid::RaftTxId;
+pub use raft_types::LogId;
+pub use raft_types::LogIndex;
+pub use raft_types::NodeId;
+pub use raft_types::Term;
 use serde::Deserialize;
 use serde::Serialize;
 
 mod errors;
 mod match_seq;
+
+mod cluster;
+mod cmd;
+mod log_entry;
+mod raft_txid;
+mod raft_types;
 
 #[cfg(test)]
 mod match_seq_test;
