+++ common/metatypes/src/raft_txid.rs
