+++ common/metatypes/src/raft_types.rs
