+++ metasrv/src/executor/kv_handlers.rs
@@ -14,6 +14,8 @@
 //
 
 use common_exception::ErrorCode;
+use common_metatypes::Cmd;
+use common_metatypes::LogEntry;
 use common_metatypes::Operation;
 use common_store_api_sdk::kv_api_impl::GetKVAction;
 use common_store_api_sdk::kv_api_impl::GetKVActionResult;
@@ -28,8 +30,6 @@ use common_store_api_sdk::kv_api_impl::UpsertKVActionResult;
 use crate::executor::action_handler::RequestHandler;
 use crate::executor::ActionHandler;
 use crate::raft::state_machine::AppliedState;
-use crate::raft::types::Cmd;
-use crate::raft::types::LogEntry;
 
 #[async_trait::async_trait]
 impl RequestHandler<UpsertKVAction> for ActionHandler {
