+++ metasrv/src/meta_service/network.rs
@@ -23,6 +23,8 @@ use async_raft::raft::InstallSnapshotResponse;
 use async_raft::raft::VoteRequest;
 use async_raft::raft::VoteResponse;
 use async_raft::RaftNetwork;
+use common_metatypes::LogEntry;
+use common_metatypes::NodeId;
 use common_tracing::tracing;
 use tonic::transport::channel::Channel;
 
@@ -30,8 +32,6 @@ use crate::meta_service::MetaRaftStore;
 use crate::meta_service::MetaServiceClient;
 use crate::meta_service::RaftMes;
 use crate::meta_service::RetryableError;
-use crate::raft::types::LogEntry;
-use crate::raft::types::NodeId;
 
 /// Impl grpc method `write`
 impl tonic::IntoRequest<RaftMes> for LogEntry {
