+++ metasrv/src/meta_service/raftmeta_test.rs
@@ -18,6 +18,8 @@ use std::sync::Arc;
 use async_raft::RaftMetrics;
 use async_raft::State;
 use common_metatypes::MatchSeq;
+use common_metatypes::NodeId;
+use common_metatypes::RaftTxId;
 use common_runtime::tokio;
 use common_runtime::tokio::time::Duration;
 use common_tracing::tracing;
@@ -30,8 +32,6 @@ use crate::meta_service::LogEntry;
 use crate::meta_service::MetaNode;
 use crate::meta_service::RetryableError;
 use crate::raft::state_machine::AppliedState;
-use crate::raft::types::NodeId;
-use crate::raft::types::RaftTxId;
 use crate::tests::assert_meta_connection;
 use crate::tests::service::new_test_context;
 use crate::tests::service::MetaSrvTestContext;
