+++ metasrv/src/raft/sled_key_spaces.rs
@@ -14,6 +14,9 @@
 
 use async_raft::raft::Entry;
 use common_metatypes::KVValue;
+use common_metatypes::LogIndex;
+use common_metatypes::Node;
+use common_metatypes::NodeId;
 use common_metatypes::SeqValue;
 
 use crate::meta_service::LogEntry;
@@ -21,9 +24,6 @@ use crate::raft::state::RaftStateKey;
 use crate::raft::state::RaftStateValue;
 use crate::raft::state_machine::StateMachineMetaKey;
 use crate::raft::state_machine::StateMachineMetaValue;
-use crate::raft::types::LogIndex;
-use crate::raft::types::Node;
-use crate::raft::types::NodeId;
 use crate::sled_store::sled_key_space::SledKeySpace;
 use crate::sled_store::SeqNum;
