+++ metasrv/src/raft/state_machine/applied_state.rs
@@ -15,6 +15,7 @@
 use async_raft::AppDataResponse;
 use common_metatypes::Database;
 use common_metatypes::KVValue;
+use common_metatypes::Node;
 use common_metatypes::SeqValue;
 use common_metatypes::Table;
 use common_store_api_sdk::storage_api_impl::DataPartInfo;
@@ -23,7 +24,6 @@ use serde::Serialize;
 
 use crate::meta_service::RaftMes;
 use crate::meta_service::RetryableError;
-use crate::raft::types::Node;
 
 /// The state of an applied raft log.
 /// Normally it includes two fields: the state before applying and the state after applying the log.
