+++ metasrv/src/raft/state_machine/placement.rs
@@ -16,9 +16,9 @@ use std::collections::hash_map::DefaultHasher;
 use std::hash::Hash;
 use std::hash::Hasher;
 
-use crate::raft::types::Node;
-use crate::raft::types::NodeId;
-use crate::raft::types::Slot;
+use common_metatypes::Node;
+use common_metatypes::NodeId;
+use common_metatypes::Slot;
 
 /// IPlacement defines the behavior of an algo to assign file to nodes.
 /// An placement algo considers the replication config, such as number of copies,
