+++ metasrv/src/raft/state_machine/sm.rs
@@ -25,9 +25,13 @@ use common_exception::ToErrorCode;
 use common_metatypes::Database;
 use common_metatypes::KVMeta;
 use common_metatypes::KVValue;
+use common_metatypes::LogId;
 use common_metatypes::MatchSeqExt;
+use common_metatypes::Node;
+use common_metatypes::NodeId;
 use common_metatypes::Operation;
 use common_metatypes::SeqValue;
+use common_metatypes::Slot;
 use common_metatypes::Table;
 use common_planners::Part;
 use common_planners::Statistics;
@@ -54,10 +58,6 @@ use crate::raft::state_machine::StateMachineMetaKey::Initialized;
 use crate::raft::state_machine::StateMachineMetaKey::LastApplied;
 use crate::raft::state_machine::StateMachineMetaKey::LastMembership;
 use crate::raft::state_machine::StateMachineMetaValue;
-use crate::raft::types::LogId;
-use crate::raft::types::Node;
-use crate::raft::types::NodeId;
-use crate::raft::types::Slot;
 use crate::sled_store::get_sled_db;
 use crate::sled_store::AsKeySpace;
 use crate::sled_store::SledSerde;
