--- metasrv/src/raft/types/mod.rs
@@ -1,29 +1,0 @@
-// Copyright 2021 Datafuse Labs.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-pub use cluster::Node;
-pub use cluster::Slot;
-pub use cmd::Cmd;
-pub use log_entry::LogEntry;
-pub use raft_txid::RaftTxId;
-pub use raft_types::LogId;
-pub use raft_types::LogIndex;
-pub use raft_types::NodeId;
-pub use raft_types::Term;
-
-mod cluster;
-mod cmd;
-mod log_entry;
-mod raft_txid;
-mod raft_types;
