+++ store/src/executor/meta_handlers.rs
@@ -22,6 +22,10 @@ use common_arrow::arrow::io::ipc::write::common::IpcWriteOptions;
 use common_arrow::arrow_flight::utils::flight_data_from_arrow_schema;
 use common_arrow::arrow_flight::FlightData;
 use common_exception::ErrorCode;
+use common_metatypes::Cmd::CreateDatabase;
+use common_metatypes::Cmd::CreateTable;
+use common_metatypes::Cmd::DropDatabase;
+use common_metatypes::Cmd::DropTable;
 use common_metatypes::Database;
 use common_metatypes::Table;
 use common_store_api_sdk::meta_api_impl::CreateDatabaseAction;
@@ -43,10 +47,6 @@ use common_store_api_sdk::meta_api_impl::GetTableExtReq;
 use log::info;
 use metasrv::meta_service::LogEntry;
 use metasrv::raft::state_machine::AppliedState;
-use metasrv::raft::types::Cmd::CreateDatabase;
-use metasrv::raft::types::Cmd::CreateTable;
-use metasrv::raft::types::Cmd::DropDatabase;
-use metasrv::raft::types::Cmd::DropTable;
 
 use crate::executor::action_handler::RequestHandler;
 use crate::executor::ActionHandler;
